change displayItem.js and move getData.js
